feat(plasma-temple): export by platforms
