+++ packages/plasma-temple/src/components/DeviceComponent/DeviceComponent.tsx
@@ -3,25 +3,27 @@ import { detectDevice } from '@salutejs/plasma-ui';
 
 import { AnyObject } from '../../types';
 
-export interface DeviceComponentProps<Props extends AnyObject = AnyObject> {
+export interface DeviceComponentProps<Props extends AnyObject = AnyObject, R = unknown> {
     sberbox: React.ComponentType<Props>;
     sberportal: React.ComponentType<Props>;
     mobile: React.ComponentType<Props>;
     props: Props;
+    innerRef?: React.Ref<R>;
 }
 
-export function DeviceComponent<Props extends AnyObject = AnyObject>({
+export function DeviceComponent<Props extends AnyObject = AnyObject, R = unknown>({
     sberbox: SberBoxComponent,
     sberportal: SberPortalComponent,
     mobile: MobileComponent,
     props,
-}: DeviceComponentProps<Props>) {
+    innerRef,
+}: DeviceComponentProps<Props, R>) {
     switch (detectDevice()) {
         case 'sberBox':
-            return <SberBoxComponent {...props} />;
+            return <SberBoxComponent {...props} ref={innerRef} />;
         case 'sberPortal':
-            return <SberPortalComponent {...props} />;
+            return <SberPortalComponent {...props} ref={innerRef} />;
         default:
-            return <MobileComponent {...props} />;
+            return <MobileComponent {...props} ref={innerRef} />;
     }
 }
