+++ packages/plasma-temple/src/components/index@sberbox.ts
@@ -5,3 +5,4 @@ export * from './Product/index@sberbox';
 export { Grid } from './Grid/Grid';
 export { HeroSliderSberbox as HeroSlider } from './HeroSlider/HeroSlider@sberbox';
 export { Card } from './Card/Card';
+export { GallerySberBox as Gallery } from './GalleryNew/Gallery@sberbox';
