+++ packages/plasma-temple/src/components/index@sberportal.ts
@@ -5,3 +5,4 @@ export * from './Product/index@sberportal';
 export { Grid } from './Grid/Grid';
 export { HeroSliderSberportal as HeroSlider } from './HeroSlider/HeroSlider@sberportal';
 export { Card } from './Card/Card';
+export { GallerySberPortal as Gallery } from './GalleryNew/Gallery@sberportal';
